Pass input to plots and user cb
